# apidocs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "\u7edf\u4e00\u9519\u8bef\u54cd\u5e94\u683c\u5f0f\uff08Pinned Error Contract\uff09\u3002\n\n\u8be5\u7ed3\u6784\u6700\u521d\u7528\u4e8e v2\uff0c\u4f46\u5728\u201c\u5408\u5e76 v2 \u5230 v1\uff08\u4ec5\u4fdd\u7559 /api/v1\uff09\u201d\u540e\uff0c\n\u670d\u52a1\u7aef\u5bf9\u5916\u7edf\u4e00\u4f7f\u7528\u6b64\u9519\u8bef\u54cd\u5e94\u683c\u5f0f\uff0c\u4fbf\u4e8e\u5ba2\u6237\u7aef\u4e0e Web \u7aef\u505a\u4e00\u81f4\u7684\u9519\u8bef\u89e3\u6790\u3002",
        "properties": {
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Details"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "SettingDeleteRequest": {
        "properties": {
          "client_updated_at_ms": {
            "default": 0,
            "title": "Client Updated At Ms",
            "type": "integer"
          }
        },
        "title": "SettingDeleteRequest",
        "type": "object"
      },
      "SettingUpsertRequest": {
        "properties": {
          "client_updated_at_ms": {
            "default": 0,
            "title": "Client Updated At Ms",
            "type": "integer"
          },
          "value_json": {
            "additionalProperties": true,
            "title": "Value Json",
            "type": "object"
          }
        },
        "title": "SettingUpsertRequest",
        "type": "object"
      },
...
      "SyncMutation": {
        "properties": {
          "client_updated_at_ms": {
            "default": 0,
            "title": "Client Updated At Ms",
            "type": "integer"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "entity_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Entity Id",
            "type": "string"
          },
          "op": {
            "enum": [
              "upsert",
              "delete"
            ],
            "title": "Op",
            "type": "string"
          },
          "resource": {
            "enum": [
              "note",
              "user_setting",
              "todo_list",
              "todo_item",
              "todo_occurrence"
            ],
            "title": "Resource",
            "type": "string"
          }
        },
        "required": [
          "resource",
          "op",
          "entity_id"
        ],
        "title": "SyncMutation",
        "type": "object"
      },
      "SyncPushRequest": {
        "properties": {
          "mutations": {
            "items": {
              "$ref": "#/components/schemas/SyncMutation"
            },
            "title": "Mutations",
            "type": "array"
          }
        },
        "title": "SyncPushRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Flow Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "operationId": "list_settings_api_v1_settings_get",
        "parameters": [
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/{key}": {
      "delete": {
        "operationId": "delete_setting_api_v1_settings__key__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Setting",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "upsert_setting_api_v1_settings__key__put",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Setting",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/v1/sync/pull": {
      "get": {
        "operationId": "pull_api_v1_sync_pull_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pull",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/v1/sync/push": {
      "post": {
        "operationId": "push_api_v1_sync_push_post",
        "parameters": [
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Push",
        "tags": [
          "sync"
        ]
      }
    },
...
  }
}